+++ TaxiTest/Models/Product.cs
@@ -14,7 +14,7 @@ namespace TaxiTest.Models
 
         public int quantity { get; set; }
 
-        public int paidPrice { get; set; }
+        public double paidPrice { get; set; }
 
         public int unitPrice { get; set; }
